# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CONVIVIUM API",
    "version": "0.1.0",
    "description": "Modo edge"
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/v1/orders/checks": {
      "get": {
        "tags": [
          "comandas"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "abierta",
                "pidio_cuenta",
                "cobrada",
                "cancelada"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "tags": [
          "comandas"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "properties": {
                      "kind": {
                        "type": "string",
                        "enum": [
                          "mesa"
                        ]
                      },
                      "tableId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "guests": {
                        "type": "integer",
                        "minimum": 1
                      }
                    },
                    "required": [
                      "kind",
                      "tableId",
                      "guests"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "kind": {
                        "type": "string",
                        "enum": [
                          "barra"
                        ]
                      },
                      "name": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "required": [
                      "kind",
                      "name"
                    ],
                    "additionalProperties": false
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/orders/checks/{id}": {
      "get": {
        "tags": [
          "comandas"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "mesa",
                        "barra"
                      ]
                    },
                    "tableId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "name": {
                      "type": "string",
                      "nullable": true
                    },
                    "guests": {
                      "type": "integer",
                      "nullable": true
                    },
                    "waiterId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "abierta",
                        "pidio_cuenta",
                        "cobrada",
                        "cancelada"
                      ]
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "checkId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "productId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "productName": {
                            "type": "string"
                          },
                          "stationId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "quantity": {
                            "type": "integer"
                          },
                          "modifiers": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "note": {
                            "type": "string",
                            "nullable": true
                          },
                          "guest": {
                            "type": "integer",
                            "nullable": true
                          },
                          "course": {
                            "type": "string",
                            "enum": [
                              "entrada",
                              "fuerte",
                              "postre",
                              "bebida",
                              "sin_tiempo"
                            ]
                          },
                          "state": {
                            "type": "string",
                            "enum": [
                              "pendiente",
                              "enviado",
                              "en_preparacion",
                              "listo",
                              "entregado",
                              "cancelado",
                              "devuelto"
                            ]
                          },
                          "unitPrice": {
                            "type": "integer",
                            "description": "Monto en centavos MXN"
                          },
                          "priority": {
                            "type": "string",
                            "enum": [
                              "normal",
                              "rehacer"
                            ]
                          },
                          "createdBy": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "sentAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "readyAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "deliveredAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "checkId",
                          "productId",
                          "productName",
                          "stationId",
                          "quantity",
                          "modifiers",
                          "note",
                          "guest",
                          "course",
                          "state",
                          "unitPrice",
                          "priority",
                          "createdBy",
                          "sentAt",
                          "readyAt",
                          "deliveredAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "subtotal": {
                      "type": "integer",
                      "description": "Monto en centavos MXN"
                    },
                    "discounts": {
                      "type": "integer",
                      "description": "Monto en centavos MXN"
                    },
                    "total": {
                      "type": "integer",
                      "description": "Monto en centavos MXN"
                    },
                    "openedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "kind",
                    "tableId",
                    "name",
                    "guests",
                    "waiterId",
                    "status",
                    "items",
                    "subtotal",
                    "discounts",
                    "total",
                    "openedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/orders/checks/{id}/items": {
      "post": {
        "tags": [
          "comandas"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "quantity": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "modifierIds": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "default": []
                        },
                        "note": {
                          "type": "string",
                          "maxLength": 140
                        },
                        "guest": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "course": {
                          "type": "string",
                          "enum": [
                            "entrada",
                            "fuerte",
                            "postre",
                            "bebida",
                            "sin_tiempo"
                          ],
                          "default": "sin_tiempo"
                        },
                        "fireNow": {
                          "type": "boolean",
                          "default": true
                        }
                      },
                      "required": [
                        "productId",
                        "quantity"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1
                  }
                },
                "required": [
                  "items"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/orders/checks/{id}/fire": {
      "post": {
        "tags": [
          "comandas"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "course": {
                    "type": "string",
                    "enum": [
                      "entrada",
                      "fuerte",
                      "postre",
                      "bebida",
                      "sin_tiempo"
                    ]
                  }
                },
                "required": [
                  "course"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/orders/checks/{id}/request-bill": {
      "post": {
        "tags": [
          "comandas"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/orders/items/move": {
      "post": {
        "tags": [
          "comandas"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "itemIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "minItems": 1
                  },
                  "toCheckId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "itemIds",
                  "toCheckId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/orders/items/{id}/transition": {
      "post": {
        "tags": [
          "comandas"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "to": {
                    "type": "string",
                    "enum": [
                      "pendiente",
                      "enviado",
                      "en_preparacion",
                      "listo",
                      "entregado",
                      "cancelado",
                      "devuelto"
                    ]
                  }
                },
                "required": [
                  "to"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/orders/items/{id}/cancel": {
      "post": {
        "tags": [
          "comandas"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reasonId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "note": {
                    "type": "string"
                  }
                },
                "required": [
                  "reasonId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/orders/items/{id}/return": {
      "post": {
        "tags": [
          "comandas"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reasonId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "resolution": {
                    "type": "string",
                    "enum": [
                      "rehacer",
                      "retirar_de_cuenta"
                    ]
                  },
                  "note": {
                    "type": "string"
                  }
                },
                "required": [
                  "reasonId",
                  "resolution"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ]
}